+++ packages/Peptides/src/utils/misc.ts
@@ -131,4 +131,5 @@ export function setGridProps(grid: DG.Grid): void {
   grid.props.showRowHeader = false;
   grid.props.showCurrentRowIndicator = false;
   grid.root.style.width = '100%';
+  grid.root.style.maxWidth = '100%';
 }