+++ libraries/compute-utils/function-views/src/pipeline-view.ts
@@ -566,10 +566,7 @@ export class PipelineView extends ComputationView {
         pulledChildRun.func.nqName === step.func.nqName);
 
       if (corrChildRun) {
-        const childRun = await historyUtils.loadRun(corrChildRun.id);
-        step.view.lastCall = deepCopy(childRun);
-        step.view.linkFunccall(childRun);
-        step.view.isHistorical.next(true);
+        await step.view.loadRun(corrChildRun.id);
 
         step.visibility.next(VISIBILITY_STATE.VISIBLE);
       } else