+++ packages/Peptides/CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Bug Fixes
 
 * Fixed WebLogo in header margin width.
+* Widgets in Context panel will now fill all the available width.
 
 ## 1.14.1 (2023-09-28)